feat: Install the axios package. Rename the locale service file.
