+++ nextjs/package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^5.16.6",
     "@mui/material": "^5.16.6",
+    "axios": "^1.7.3",
     "next": "14.2.5",
     "next-intl": "^3.17.2",
     "react": "^18",
